agentic-commerce-system/requirements.txt
# Core LangChain libraries
langchain==0.1.20
langgraph==0.0.48
langchain-openai==0.1.6

# For managing environment variables
python-dotenv==1.0.1